feat: results chat support for attachments
